backend/requirements.txt
# Core dependencies
crewai>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0

# API Framework
fastapi>=0.110.0
uvicorn>=0.27.0
websockets>=12.0
python-multipart>=0.0.9

# GitHub Integration
PyGithub>=2.1.1
gitpython>=3.1.40

# Utilities
pydantic>=2.5.0
httpx>=0.26.0
aiofiles>=23.2.1

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.1.0
flake8>=7.0.0